fixing mini bugs
