+++ src/components/ModalForm.jsx
@@ -1,12 +1,9 @@
-// src/components/ModalForm.jsx
 import { useState, useEffect } from "react";
 import { motion } from "framer-motion";
 import { formConfig } from "../constants/formConfig";
-import { Switch } from "@/components/ui/switch";
-import { Label } from "@/components/ui/label";
-import { ImagePlus } from "lucide-react";
+import FormFields from "./FormFields";
 
-// Servicios API
+// Servicios API (mismas rutas que tenías, ajustadas un nivel más arriba)
 import {
   getSocialMedia,
   updateSocialMedia,
@@ -41,7 +38,7 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
   const [machineTypes, setMachineTypes] = useState([]);
   const [serviceTypes, setServiceTypes] = useState([]);
 
-  // === Cargar listas de tipos según sección + intentar mapear typeName->id en edición ===
+  // === Cargar listas de tipos según sección + mapear typeName->id en edición ===
   useEffect(() => {
     const mapTypeNameIfNeeded = (data) => {
       if (item?.typeName) {
@@ -96,7 +93,7 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
     }
   }, [isSocial, item]);
 
-  // === Prellenado general para edición/creación (excepto sociales que se cargan arriba) ===
+  // === Prellenado general (excepto sociales) ===
   useEffect(() => {
     if (item && !isSocial) {
       setForm({
@@ -107,7 +104,6 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
         descripcion: item.description || "",
         imagen: item.imageUrl || "",
         file: null,
-        // ✅ mapear "destacado" UI <-> isFeatured (DB)
         destacado:
           item.outstanding === "si" ||
           item.destacado === true ||
@@ -134,7 +130,7 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
     setForm((prev) => ({ ...prev, [e.target.name]: e.target.value }));
 
   const handleFileChange = (e) => {
-    const file = e.target.files[0];
+    const file = e.target.files?.[0];
     setForm((prev) => ({ ...prev, file }));
     if (file) {
       const reader = new FileReader();
@@ -150,11 +146,12 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
       // BANNER
       if (isBanner) {
         if (!form.nombre?.trim()) return alert("Completa el nombre del banner");
-        if (!form.file && !item)
-          return alert("Selecciona una imagen para el banner");
+        if (!form.file && !item) return alert("Selecciona una imagen para el banner");
+
         const formData = new FormData();
         formData.append("name", form.nombre);
         if (form.file) formData.append("image", form.file);
+
         if (item) {
           await updateBanner(form.id, formData);
         } else {
@@ -164,7 +161,7 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
         return onClose();
       }
 
-      // SOCIALES (editar url por tipo/enum)
+      // SOCIALES
       if (isSocial) {
         if (!form.url?.trim()) return alert("La URL es obligatoria");
         await updateSocialMedia(form.tipo, { url: form.url });
@@ -208,17 +205,17 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
 
       // PRODUCTOS
       if (type === "productos") {
-        if (!form.nombre?.trim())
-          return alert("Completa el nombre del producto");
+        if (!form.nombre?.trim()) return alert("Completa el nombre del producto");
         if (!form.tipo) return alert("Selecciona un tipo de producto");
+
         const formData = new FormData();
-        if (!item || form.nombre !== item.name)
-          formData.append("name", form.nombre);
+        if (!item || form.nombre !== item.name) formData.append("name", form.nombre);
         formData.append("typeId", form.tipo);
         formData.append("capacity", parseFloat(form.capacidad) || 0);
         formData.append("description", form.descripcion || "");
         formData.append("isFeatured", form.isFeatured ? "true" : "false");
         if (form.file) formData.append("image", form.file);
+
         if (item) {
           await updateProduct(form.id, formData);
         } else {
@@ -230,16 +227,16 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
 
       // MAQUINARIA
       if (type === "maquinaria") {
-        if (!form.nombre?.trim())
-          return alert("Completa el nombre de la maquinaria");
+        if (!form.nombre?.trim()) return alert("Completa el nombre de la maquinaria");
         if (!form.tipo) return alert("Selecciona un tipo de maquinaria");
+
         const formData = new FormData();
-        if (!item || form.nombre !== item.name)
-          formData.append("name", form.nombre);
+        if (!item || form.nombre !== item.name) formData.append("name", form.nombre);
         formData.append("typeId", form.tipo);
         formData.append("description", form.descripcion || "");
         formData.append("isFeatured", form.isFeatured ? "true" : "false");
         if (form.file) formData.append("image", form.file);
+
         if (item) {
           await updateMachine(form.id, formData);
         } else {
@@ -251,16 +248,16 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
 
       // SERVICIOS
       if (type === "servicios") {
-        if (!form.nombre?.trim())
-          return alert("Completa el nombre del servicio");
+        if (!form.nombre?.trim()) return alert("Completa el nombre del servicio");
         if (!form.tipo) return alert("Selecciona un tipo de servicio");
+
         const formData = new FormData();
-        if (!item || form.nombre !== item.name)
-          formData.append("name", form.nombre);
+        if (!item || form.nombre !== item.name) formData.append("name", form.nombre);
         formData.append("typeId", form.tipo);
         formData.append("description", form.descripcion || "");
         formData.append("isFeatured", form.isFeatured ? "true" : "false");
         if (form.file) formData.append("image", form.file);
+
         if (item) {
           await updateService(form.id, formData);
         } else {
@@ -275,10 +272,9 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
     }
   };
 
-  // === Campos dinámicos ===
+  // === Campos dinámicos (idéntico a tu componente original) ===
   const normalizedType = type.charAt(0).toUpperCase() + type.slice(1);
 
-  // Fallback explícito para formularios de "Tipos" (asegura que siempre muestre el input)
   const typeLabels = {
     ProductosTipos: "Nombre del tipo de producto",
     ServiciosTipos: "Nombre del tipo de servicio",
@@ -307,144 +303,19 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
           {item ? "Editar" : "Registrar"} {type}
         </h2>
 
-        {/* Campos dinámicos */}
-        <div className="space-y-4">
-          {fields.map((field) => {
-            const id = `field-${field.name}`;
+        <FormFields
+          fields={fields}
+          form={form}
+          setForm={setForm}
+          handleChange={handleChange}
+          handleFileChange={handleFileChange}
+          options={{ productTypes, machineTypes, serviceTypes }}
+        />
 
-            if (["text", "number", "url"].includes(field.type)) {
-              return (
-                <div
-                  key={field.name}
-                  className="space-y-1 hover:scale-[1.02] transition-transform"
-                >
-                  <Label htmlFor={id} className="text-gray-700">
-                    {field.label}
-                  </Label>
-                  <input
-                    id={id}
-                    name={field.name}
-                    type={field.type}
-                    value={form[field.name] ?? ""}
-                    onChange={handleChange}
-                    placeholder={field.placeholder || ""}
-                    readOnly={!!field.readonly}
-                    className="w-full rounded-lg bg-gray-50 border border-gray-300 text-gray-900 px-3 py-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
-                  />
-                </div>
-              );
-            }
-
-            if (field.type === "textarea") {
-              return (
-                <div
-                  key={field.name}
-                  className="space-y-1 hover:scale-[1.02] transition-transform"
-                >
-                  <Label htmlFor={id} className="text-gray-700">
-                    {field.label}
-                  </Label>
-                  <textarea
-                    id={id}
-                    name={field.name}
-                    value={form[field.name] ?? ""}
-                    onChange={handleChange}
-                    placeholder={field.placeholder || ""}
-                    rows={3}
-                    className="w-full rounded-lg bg-gray-50 border border-gray-300 text-gray-900 px-3 py-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
-                  />
-                </div>
-              );
-            }
-
-            if (field.type === "select") {
-              const options =
-                field.optionsSource === "productTypes"
-                  ? productTypes
-                  : field.optionsSource === "machineTypes"
-                    ? machineTypes
-                    : field.optionsSource === "serviceTypes"
-                      ? serviceTypes
-                      : serviceTypes;
-
-              return (
-                <div
-                  key={field.name}
-                  className="space-y-1 hover:scale-[1.02] transition-transform"
-                >
-                  <Label htmlFor={id} className="text-gray-700">
-                    {field.label}
-                  </Label>
-                  <select
-                    id={id}
-                    name={field.name}
-                    value={form[field.name] ?? ""}
-                    onChange={handleChange}
-                    className="w-full rounded-lg bg-gray-50 border border-gray-300 text-gray-900 px-3 py-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
-                  >
-                    <option value="">{`Seleccione ${field.label}`}</option>
-                    {options?.map((opt) => (
-                      <option key={opt.id} value={opt.id}>
-                        {opt.name}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-              );
-            }
-
-            if (field.type === "image") {
-              return (
-                <UploadImage
-                  key={field.name}
-                  form={form}
-                  handleFileChange={handleFileChange}
-                  inputId={`fileInput-${field.name}`}
-                  label={field.label}
-                />
-              );
-            }
-
-            if (field.type === "switch") {
-              // ✅ El switch maneja `destacado` en UI pero actualiza y lee `isFeatured`
-              const isFeaturedValue =
-                field.name === "destacado"
-                  ? !!form.isFeatured
-                  : !!form[field.name];
-
-              return (
-                <div
-                  key={field.name}
-                  className="flex items-center justify-between pt-1 hover:scale-[1.02] transition-transform"
-                >
-                  <Label htmlFor={id} className="text-gray-700 font-medium">
-                    {field.label}
-                  </Label>
-                  <Switch
-                    id={id}
-                    checked={isFeaturedValue}
-                    onCheckedChange={(val) =>
-                      setForm((prev) => ({
-                        ...prev,
-                        [field.name]: val,
-                        ...(field.name === "destacado" ? { isFeatured: val } : {}),
-                      }))
-                    }
-                    className="data-[state=checked]:bg-gradient-to-r from-cyan-400 to-blue-600"
-                  />
-                </div>
-              );
-            }
-
-            return null;
-          })}
-        </div>
-
-        {/* Botones */}
         <div className="flex justify-end gap-3 pt-6">
           <button
             onClick={onClose}
-            className="px-4 py-2 rounded-lg bg-gray-300 text-gray-800 font-medium hover:bg-gray-200 hover:scale-110 transition shadow"
+            className="px-4 py-2 rounded-lg bg-gray-200 text-gray-800 font-medium hover:bg-gray-300 hover:scale-110 transition shadow"
           >
             Cancelar
           </button>
@@ -459,49 +330,5 @@ const ModalForm = ({ type, onClose, onSave, item, lastId }) => {
     </div>
   );
 };
-
-// === Upload de imagen ===
-const UploadImage = ({ form, handleFileChange, inputId, label }) => (
-  <div className="space-y-2 hover:scale-[1.02] transition-transform">
-    {label && <Label className="text-gray-700">{label}</Label>}
-    <div
-      className="border-2 border-dashed border-gray-300 bg-gray-50 p-4 rounded-lg cursor-pointer hover:bg-gray-100 transition flex flex-col items-center justify-center text-center"
-      onDragOver={(e) => e.preventDefault()}
-      onDrop={(e) => {
-        e.preventDefault();
-        const file = e.dataTransfer.files[0];
-        if (file) handleFileChange({ target: { files: [file] } });
-      }}
-    >
-      <div className="flex flex-col items-center justify-center">
-        <ImagePlus className="w-10 h-10 text-gray-400 mb-2" />
-        <p className="text-gray-500 text-sm">
-          Arrastra una imagen aquí o haz clic para seleccionarla
-        </p>
-      </div>
-      <input
-        type="file"
-        accept="image/*"
-        onChange={handleFileChange}
-        className="hidden"
-        id={inputId}
-      />
-      <label
-        htmlFor={inputId}
-        className="mt-3 inline-block px-3 py-1 rounded-lg bg-gradient-to-r from-cyan-400 to-blue-500 text-white cursor-pointer shadow hover:shadow-lg hover:opacity-90 transition"
-      >
-        Seleccionar archivo
-      </label>
-    </div>
-
-    {form.imagen && (
-      <img
-        src={form.file ? URL.createObjectURL(form.file) : form.imagen}
-        alt="Vista previa"
-        className="w-full h-40 object-cover rounded-lg transition-transform hover:scale-[1.02]"
-      />
-    )}
-  </div>
-);
 
 export default ModalForm;
